+++ TombEngine/Game/items.cpp
@@ -411,6 +411,9 @@ void KillItem(short const itemNumber)
 		if (item == Lara.TargetEntity)
 			Lara.TargetEntity = nullptr;
 
+		if (item == Lara.SecondaryTargetEntity)
+			Lara.SecondaryTargetEntity = nullptr;
+
 		// AI target generation uses a hack with making a dummy item without ObjectNumber.
 		// Therefore, a check should be done here to prevent access violation.
 		if (item->ObjectNumber != GAME_OBJECT_ID::ID_NO_OBJECT && item->IsBridge())