+++ TombEngine/Scripting/Internal/TEN/Flow/Settings/Settings.cpp
@@ -30,15 +30,15 @@ namespace TEN::Scripting
 	
 		// NOTE: Since Weapons array is bound to Lua directly and Lua accesses this array by native enum, where 0 is NONE, and 1 is PISTOLS,
 		// 0 index is omitted due to Lua indexing arrays starting from 1. 1 must be subtracted from initializer index.
-		Weapons[(int)LaraWeaponType::Pistol          - 1] = { 8.0f,  BLOCK(8),  9,  (int)BLOCK(0.65f), 1,  1,  30, ScriptColor(192, 128, 0), 9,  3, true,  true,  true,  false, false, Vec3(  0, 120, 30)  };
-		Weapons[(int)LaraWeaponType::Revolver        - 1] = { 4.0f,  BLOCK(8),  16, (int)BLOCK(0.65f), 21, 21, 6,  ScriptColor(192, 128, 0), 9,  3, true,  false, true,  false, false, Vec3(-10, 130, 65)  };
-		Weapons[(int)LaraWeaponType::Uzi             - 1] = { 8.0f,  BLOCK(8),  3,  (int)BLOCK(0.65f), 1,  1,  30, ScriptColor(192, 128, 0), 9,  2, true,  true,  true,  false, false, Vec3(  0, 110, 40)  };
-		Weapons[(int)LaraWeaponType::Shotgun         - 1] = { 10.0f, BLOCK(8),  0,  (int)BLOCK(0.50f), 3,  3,  6,  ScriptColor(192, 128, 0), 12, 3, true,  true,  false, false, false, Vec3(  0, 210, 40)  };
-		Weapons[(int)LaraWeaponType::HK              - 1] = { 4.0f,  BLOCK(12), 0,  (int)BLOCK(0.50f), 4,  4,  30, ScriptColor(192, 128, 0), 12, 2, true,  true,  true,  false, false, Vec3(  0, 220, 102) };
-		Weapons[(int)LaraWeaponType::Crossbow        - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 5,  20, 10, ScriptColor(192, 128, 0), 0,  0, false, false, false, false, false, Vec3(  0, 240, 50)  };
-		Weapons[(int)LaraWeaponType::GrenadeLauncher - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 30, 30, 10, ScriptColor(192, 128, 0), 0,  0, true,  false, false, false, false, Vec3(  0, 190, 50)  };
-		Weapons[(int)LaraWeaponType::RocketLauncher  - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 30, 30, 1,  ScriptColor(192, 128, 0), 0,  0, true,  false, false, false, false, Vec3(  0, 90,  90)  };
-		Weapons[(int)LaraWeaponType::HarpoonGun      - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 6,  6,  10, ScriptColor(192, 128, 0), 0,  0, false, false, false, false, false, Vec3(-15, 240, 50)  };
+		Weapons[(int)LaraWeaponType::Pistol          - 1] = { 8.0f,  BLOCK(8),  9,  (int)BLOCK(0.65f), 1,  1,  30, ScriptColor(192, 128, 0), 9,  3, true,  true,  true,  false, false, Vec3(  0, 120, 30), true };
+		Weapons[(int)LaraWeaponType::Revolver        - 1] = { 4.0f,  BLOCK(8),  16, (int)BLOCK(0.65f), 21, 21, 6,  ScriptColor(192, 128, 0), 9,  3, true,  false, true,  false, false, Vec3(-10, 130, 65), false };
+		Weapons[(int)LaraWeaponType::Uzi             - 1] = { 8.0f,  BLOCK(8),  3,  (int)BLOCK(0.65f), 1,  1,  30, ScriptColor(192, 128, 0), 9,  2, true,  true,  true,  false, false, Vec3(  0, 110, 40), true};
+		Weapons[(int)LaraWeaponType::Shotgun         - 1] = { 10.0f, BLOCK(8),  0,  (int)BLOCK(0.50f), 3,  3,  6,  ScriptColor(192, 128, 0), 12, 3, true,  true,  false, false, false, Vec3(  0, 210, 40), false };
+		Weapons[(int)LaraWeaponType::HK              - 1] = { 4.0f,  BLOCK(12), 0,  (int)BLOCK(0.50f), 4,  4,  30, ScriptColor(192, 128, 0), 12, 2, true,  true,  true,  false, false, Vec3(  0, 220, 102), false };
+		Weapons[(int)LaraWeaponType::Crossbow        - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 5,  20, 10, ScriptColor(192, 128, 0), 0,  0, false, false, false, false, false, Vec3(  0, 240, 50), false };
+		Weapons[(int)LaraWeaponType::GrenadeLauncher - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 30, 30, 10, ScriptColor(192, 128, 0), 0,  0, true,  false, false, false, false, Vec3(  0, 190, 50), false };
+		Weapons[(int)LaraWeaponType::RocketLauncher  - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 30, 30, 1,  ScriptColor(192, 128, 0), 0,  0, true,  false, false, false, false, Vec3(  0, 90,  90), false };
+		Weapons[(int)LaraWeaponType::HarpoonGun      - 1] = { 8.0f,  BLOCK(8),  0,  (int)BLOCK(0.50f), 6,  6,  10, ScriptColor(192, 128, 0), 0,  0, false, false, false, false, false, Vec3(-15, 240, 50), false };
 	}
 
 	void Settings::Register(sol::table& parent)
@@ -262,7 +262,11 @@ namespace TEN::Scripting
 
 		/// Enable target occlusion by moveables and static meshes.
 		// @tfield[opt=true] bool targetObjectOcclusion If enabled, player won't be able to target enemies through moveables and static meshes.
-		"targetObjectOcclusion", &GameplaySettings::TargetObjectOcclusion);
+		"targetObjectOcclusion", &GameplaySettings::TargetObjectOcclusion,
+
+		/// Enable dual targeting for double-handed weapons.
+		// @tfield[opt=true] bool dualTarget If enabled, each arm of a double-handed weapon can independently target a separate enemy.
+		"dualTarget", &GameplaySettings::DualTarget);
 	}
 
 	/// Effects
@@ -717,6 +721,10 @@ namespace TEN::Scripting
 		/// Gunflash duration.
 		// @tfield int flashDuration specifies the duration of a gunflash effect.
 		"flashDuration", &WeaponSettings::FlashDuration,
+
+		/// Double-handed grip.
+		// @tfield[opt=false] bool doubleHanded If set to true, weapon uses both hands. Applicable only for pistols, uzis, and revolver.
+		"doubleHanded", &WeaponSettings::DoubleHanded,
 
 		/// Gun smoke.
 		// @tfield bool smoke if set to true, indicates that weapon emits gun smoke.