+++ CHANGELOG.md
@@ -89,6 +89,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 
 * Fixed `Timer` class not working correctly with single frame intervals.
 * Fixed alpha value being ignored for `Flow.Settings.UI.shadowTextColor` parameter.
+* Added `settings.Weapons[WeaponType.<...>].doubleHanded` parameter for Pistols, Uzis and Revolver to allow single or dual weapons.
 
 ## [Version 1.11]